Move code generators to separate namespace
